pyproject.toml
[project]
name = "bookingapp"
version = "0.1.0"
description = "Booking App"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "psycopg2-binary",
    "python-dotenv",
    "drf-spectacular",
    "pytest",
    "pytest-django",
    "pytest-cov",
]

[tool.uv]
dev-dependencies = [
    "pre-commit>=3.5.0",
    "ruff>=0.1.6",
]

[tool.ruff]
exclude = [
    ".git",
    "__pycache__",
    "*/migrations/*",
    ".venv",
    "venv",
    "build",
    "dist",
    ".env",
]

line-length = 88
target-version = "py312"

select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "DJ",  # flake8-django
    "RUF", # Ruff-specific rules
]

ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

fix = true

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"test_*.py" = ["S101", "PLR2004"]
"tests.py" = ["S101", "PLR2004"]
"src/*/settings/*.py" = ["F405", "F403"]  # разрешаем * imports в settings

[tool.ruff.isort]
known-third-party = ["django", "rest_framework", "psycopg2", "dotenv"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
pythonpath = ["src"]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = ["--tb=short"]
